requirements.txt
aiohttp
python-dotenv
fastapi
uvicorn
ta-lib
schedule
MetaTrader5
requests
pydantic
